genesisBlockHash as Network member
