+++ src/networks.d.ts
@@ -1,8 +1,11 @@
+/// <reference types="node" />
 import { Network as BitcoinJSNetwork } from 'ecpair/src/networks';
+export declare type GenesisBlockHash = Buffer;
 export interface Network extends BitcoinJSNetwork {
     blech32: string;
     assetHash: string;
     confidentialPrefix: number;
+    genesisBlockHash: GenesisBlockHash;
 }
 export declare const liquid: Network;
 export declare const regtest: Network;
