+++ ts_src/networks.ts
@@ -1,11 +1,29 @@
 import { Network as BitcoinJSNetwork } from 'ecpair/src/networks';
 
+export type GenesisBlockHash = Buffer;
+
+const strToGenesisHash = (str: string): GenesisBlockHash =>
+  Buffer.from(str, 'hex').reverse();
+
+const RegtestGenesisBlockHash = strToGenesisHash(
+  '00902a6b70c2ca83b5d9c815d96a0e2f4202179316970d14ea1847dae5b1ca21',
+);
+
+const TestnetGenesisBlockHash = strToGenesisHash(
+  'a771da8e52ee6ad581ed1e9a99825e5b3b7992225534eaa2ae23244fe26ab1c1',
+);
+
+const LiquidGenesisBlockHash = strToGenesisHash(
+  '1466275836220db2944ca059a3a10ef6fd2ea684b0688d2c379296888a206003',
+);
+
 // https://en.bitcoin.it/wiki/List_of_address_prefixes
 // Dogecoin BIP32 is a proposed standard: https://bitcointalk.org/index.php?topic=409731
 export interface Network extends BitcoinJSNetwork {
   blech32: string;
   assetHash: string;
   confidentialPrefix: number;
+  genesisBlockHash: GenesisBlockHash;
 }
 
 export const liquid: Network = {
@@ -21,6 +39,7 @@ export const liquid: Network = {
   wif: 0x80,
   confidentialPrefix: 12,
   assetHash: '6f0279e9ed041c3d710a9f57d0c02928416460c4b722ae3457a11eec381c526d',
+  genesisBlockHash: LiquidGenesisBlockHash,
 };
 
 export const regtest: Network = {
@@ -36,6 +55,7 @@ export const regtest: Network = {
   wif: 0xef,
   confidentialPrefix: 4,
   assetHash: '5ac9f65c0efcc4775e0baec4ec03abdde22473cd3cf33c0419ca290e0751b225',
+  genesisBlockHash: RegtestGenesisBlockHash,
 };
 
 export const testnet: Network = {
@@ -46,4 +66,5 @@ export const testnet: Network = {
   scriptHash: 19,
   confidentialPrefix: 23,
   assetHash: '144c654344aa716d6f3abcc1ca90e5641e4e2a7f633bc09fe3baf64585819a49',
+  genesisBlockHash: TestnetGenesisBlockHash,
 };
